# packages/contracts/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "PathwayOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name"
    ],
    "title": "PathwayOut",
    "type": "object"
   },
...
   "ProgramOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "institution": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Institution"
     },
     "is_synthetic": {
      "title": "Is Synthetic",
      "type": "boolean"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "parent_degree": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Degree"
     }
    },
    "required": [
     "id",
     "code",
     "name",
     "institution",
     "parent_degree",
     "description",
     "is_synthetic"
    ],
    "title": "ProgramOut",
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Curriculum decision support. Not a predictor of learning gains or an accreditation certifier.",
  "title": "Curriculum Flight Simulator API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/entities/{entity_id}/versions": {
   "get": {
    "description": "Stable identity across curriculum versions: which exact revision each version uses.",
    "operationId": "entity_across_versions_api_v1_entities__entity_id__versions_get",
    "parameters": [
     {
      "in": "path",
      "name": "entity_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Entity Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Entity Across Versions Api V1 Entities  Entity Id  Versions Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Entity Across Versions",
    "tags": [
     "entities"
    ]
   }
  },
  "/api/v1/evidence/batch": {
   "post": {
    "operationId": "get_spans_api_v1_evidence_batch_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "items": {
         "format": "uuid",
         "type": "string"
        },
        "title": "Ids",
        "type": "array"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SpanOut"
         },
         "title": "Response Get Spans Api V1 Evidence Batch Post",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Spans",
    "tags": [
     "evidence"
    ]
   }
  },
  "/api/v1/evidence/{span_id}": {
   "get": {
    "operationId": "get_span_api_v1_evidence__span_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "span_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Span Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SpanOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Span",
    "tags": [
     "evidence"
    ]
   }
  },
...
  "/api/v1/programs": {
   "get": {
    "operationId": "list_programs_api_v1_programs_get",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ProgramOut"
         },
         "title": "Response List Programs Api V1 Programs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Programs",
    "tags": [
     "programs"
    ]
   },
   "post": {
    "operationId": "create_program_api_v1_programs_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProgramIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Program",
    "tags": [
     "ai",
     "programs"
    ]
   }
  },
  "/api/v1/programs/{program_id}": {
   "get": {
    "operationId": "get_program_api_v1_programs__program_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "program_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Program Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProgramOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Program",
    "tags": [
     "programs"
    ]
   }
  },
  "/api/v1/programs/{program_id}/pathways": {
   "get": {
    "operationId": "list_pathways_api_v1_programs__program_id__pathways_get",
    "parameters": [
     {
      "in": "path",
      "name": "program_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Program Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/PathwayOut"
         },
         "title": "Response List Pathways Api V1 Programs  Program Id  Pathways Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Pathways",
    "tags": [
     "programs"
    ]
   }
  },
  "/api/v1/programs/{program_id}/versions": {
   "get": {
    "operationId": "list_versions_api_v1_programs__program_id__versions_get",
    "parameters": [
     {
      "in": "path",
      "name": "program_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Program Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/VersionOut"
         },
         "title": "Response List Versions Api V1 Programs  Program Id  Versions Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Versions",
    "tags": [
     "programs"
    ]
   },
   "post": {
    "operationId": "create_version_api_v1_programs__program_id__versions_post",
    "parameters": [
     {
      "in": "path",
      "name": "program_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Program Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VersionCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VersionOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Version",
    "tags": [
     "programs"
    ]
   }
  },
...
  "/api/v1/versions/{version_id}": {
   "get": {
    "operationId": "get_version_api_v1_versions__version_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "version_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Version Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VersionOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Version",
    "tags": [
     "versions"
    ]
   }
  },
...
  "/api/v1/versions/{version_id}/table": {
   "get": {
    "description": "Accessible list alternative to the graph: every course with its placement and direct relationships.",
    "operationId": "get_table_api_v1_versions__version_id__table_get",
    "parameters": [
     {
      "in": "path",
      "name": "version_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Version Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Table Api V1 Versions  Version Id  Table Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Table",
    "tags": [
     "graph"
    ]
   }
  },
...
 }
}